services/vanna/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
groq==0.13.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
pydantic==2.10.5
